Added confirmation alert pop-up for delete task

+++ src/CompletedTask.js
@@ -4,9 +4,12 @@ import moment from 'moment';
 class CompletedTask extends React.Component {
 
     deleteClicked = () => {
-        this.props.deleteTaskFunc(this.props.item.taskID)
+        let deleteConfirmation = window.confirm('Are you sure you want to delete the task? This action cannot be undone.');
+        if (deleteConfirmation == true) {
+            this.props.deleteTaskFunc(this.props.item.taskID)
+        }
     }
-
+    
     undoClicked = () => {
         this.props.undoClickedFunc(this.props.item.taskID)
     }
